Test function refactoring
SSDController 와 SSD, CommandBuffer 통합 테스트를 위해
Test*.cpp에서 정의된 일반 함수들을
TestCommon.h/.cpp로 이동
