+++ TeamBest_SSD/TestCommandBuffer.cpp
@@ -1,4 +1,5 @@
 ﻿#include "gmock/gmock.h"
+#include "TestCommons.h"
 #include "command_buffer.h"
 #include <iostream>
 #include <filesystem>
@@ -11,63 +12,10 @@
 using namespace testing;
 namespace fs = std::filesystem;
 
-
-int countFilesInBuffer() {
-	int count = 0;
-	for (const auto& entry : std::filesystem::directory_iterator("buffer")) {
-		if (std::filesystem::is_regular_file(entry)) {
-			++count;
-		}
-	}
-	return count;
-}
-
-std::vector<std::string> GetFilesInDirectory(const std::string& directoryPath) {
-	std::vector<std::string> fileList;
-
-	try {
-		if (fs::exists(directoryPath) && fs::is_directory(directoryPath)) {
-			for (const auto& entry : fs::directory_iterator(directoryPath)) {
-				if (fs::is_regular_file(entry.status())) {
-					fileList.push_back(entry.path().filename().string());
-				}
-			}
-		}
-	}
-	catch (const fs::filesystem_error& e) {
-#ifdef _DEBUG
-		std::cerr << "파일 목록 검사 중 오류 발생: " << e.what() << '\n';
-#endif
-	}
-
-	return fileList;
-
-}
-
-void RemoveDirectoryAndRecreate( const std::string& dirPath ) {
-	if (std::filesystem::exists(dirPath))
-		std::filesystem::remove_all(dirPath);
-
-	if (!fs::exists(dirPath)) {
-		fs::create_directory(dirPath);
-	}
-}
-
-void MakeBufferFiles(
-	const std::vector<std::string>& bufferNames, 
-	const std::string & bufferDirPath) {
-	for (const auto& bufferName : bufferNames) {
-		std::ofstream outFile(bufferDirPath + "/" + bufferName);
-		outFile.close();
-	}
-}
-
-
 TEST(TestCommandBuffer, ContructorTest) {
 	EXPECT_NO_THROW(std::shared_ptr<CommandBuffer> buffer 
 		= std::make_shared<CommandBuffer>());
 }
-
 
 TEST(TestCommandBuffer, BufferExistsAfterInit){
 	const std::string dirName = "buffer";
